Rework the exponent calculation logic, pre-echo and NF improvements (#1)
* feat: Implement normalized spectral coeffs, rework NF, SNR improvements

* feat: improve tns gate

* feat: transient handling improvements

* fix: rm debug

* feat: claen up, update the design doc, rescore visqol / zim in readme

* fix: tests

* feat: Further update readme, regenerate tables

* fix: remove sdkconfig.old from benchmark file

* fix: Add missing dr_wav.h to the repo

+++ include/pcm.h
@@ -10,21 +10,35 @@
 extern "C" {
 #endif
 
-// Load one interleaved PCM sample as Q31
+/**
+ * @brief Load one interleaved PCM sample as Q31.
+ *
+ * @param base PCM buffer base pointer.
+ * @param fmt PCM sample format.
+ * @param idx Interleaved sample index.
+ * @return Sample converted to Q31.
+ */
 static inline int32_t pcm_load_q31(const uint8_t *base, hqlc_pcm_format fmt, int idx) {
   if (fmt == HQLC_PCM16) {
     const int16_t *p = (const int16_t *)base;
-    return (int32_t)p[idx] << 16;
+    return (int32_t)((uint32_t)p[idx] << 16);
   }
   const uint8_t *p = base + 3 * idx;
   int32_t v = (int32_t)p[0] | ((int32_t)p[1] << 8) | ((int32_t)p[2] << 16);
   if (v & 0x800000) {
     v |= (int32_t)0xFF000000;
   }
-  return v << 8;
+  return (int32_t)((uint32_t)v << 8);
 }
 
-// Load one interleaved PCM sample in native range (no Q31 shift)
+/**
+ * @brief Load one interleaved PCM sample without Q31 scaling.
+ *
+ * @param base PCM buffer base pointer.
+ * @param fmt PCM sample format.
+ * @param idx Interleaved sample index.
+ * @return Sample in its native signed integer range.
+ */
 static inline int32_t pcm_load_native(const uint8_t *base, hqlc_pcm_format fmt, int idx) {
   if (fmt == HQLC_PCM16) {
     const int16_t *p = (const int16_t *)base;
@@ -38,7 +52,17 @@ static inline int32_t pcm_load_native(const uint8_t *base, hqlc_pcm_format fmt, 
   return v;
 }
 
-// Store a Q31 value to interleaved PCM with rounding and clamping
+/**
+ * @brief Store a Q31 value as interleaved PCM.
+ *
+ * The stored value is rounded to the target format. PCM16 output is clamped to
+ * the signed 16-bit range.
+ *
+ * @param base PCM buffer base pointer.
+ * @param fmt PCM sample format.
+ * @param idx Interleaved sample index.
+ * @param val_q31 Q31 sample value to store.
+ */
 static inline void
 pcm_store_q31(uint8_t *base, hqlc_pcm_format fmt, int idx, int32_t val_q31) {
   if (fmt == HQLC_PCM16) {
@@ -60,7 +84,12 @@ pcm_store_q31(uint8_t *base, hqlc_pcm_format fmt, int idx, int32_t val_q31) {
   }
 }
 
-// Clamp an int32 to int16 range
+/**
+ * @brief Clamp a 32-bit integer to the signed 16-bit range.
+ *
+ * @param x Value to clamp.
+ * @return `x` clamped to INT16_MIN..INT16_MAX.
+ */
 static inline int16_t pcm_clamp_i16(int32_t x) {
   if (x > 32767) {
     return 32767;
